requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.1
async-timeout==4.0.1
attrs==21.2.0
blessed==1.19.0
botocore==1.23.21
certifi==2021.10.8
chalice==1.26.2
charset-normalizer==2.0.9
click==8.0.3
frozenlist==1.2.0
idna==3.3
inquirer==2.8.0
jmespath==0.10.0
multidict==5.2.0
mygeotab==0.8.7
mypy-extensions==0.4.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
python-rapidjson==1.5
pytz==2021.3
PyYAML==6.0
readchar==2.0.1
requests==2.26.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
yarl==1.7.2
